# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Uniform error body, shared by every endpoint.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode",
            "description": "Machine-readable code — clients branch on this, not on `message`."
          },
          "details": {
            "description": "Optional structured payload (field path, offending value, etc.). Shape is error-specific; carried as untyped JSON so the API crate does not need to enumerate every possibility.",
            "nullable": true
          },
          "message": {
            "description": "Human-readable, user-facing explanation.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorCode": {
        "description": "Machine-readable error taxonomy.\n\nAny variant added here must also be assigned an HTTP status in [`http_status`]. The compiler enforces exhaustiveness.",
        "oneOf": [
          {
            "description": "The `game` field in a request did not name a registered game.",
            "enum": [
              "UnknownGame"
            ],
            "type": "string"
          },
          {
            "description": "An `agents[_]` entry did not name a registered agent kind.",
            "enum": [
              "UnknownAgent"
            ],
            "type": "string"
          },
          {
            "description": "The `config` blob failed validation against the target game's config schema.",
            "enum": [
              "InvalidConfig"
            ],
            "type": "string"
          },
          {
            "description": "The requested run id was not found (or has been evicted).",
            "enum": [
              "RunNotFound"
            ],
            "type": "string"
          },
          {
            "description": "The requested game id / log file was not found.",
            "enum": [
              "GameNotFound"
            ],
            "type": "string"
          },
          {
            "description": "An `offset` or `limit` query parameter was negative, zero where positive is required, or exceeds server-enforced ceilings.",
            "enum": [
              "InvalidPaginationParams"
            ],
            "type": "string"
          },
          {
            "description": "An unexpected server-side failure. The `message` is safe to show users; `details` (if populated) is meant for operators.",
            "enum": [
              "Internal"
            ],
            "type": "string"
          },
          {
            "description": "The endpoint is scaffolded but not yet implemented. Used for routes shipped as stubs ahead of their full implementation (e.g. `/api/reports` in Unit 17, proper implementation deferred to a later unit).",
            "enum": [
              "NotImplemented"
            ],
            "type": "string"
          },
          {
            "description": "`POST .../actions` was submitted with a `prompt_id` that does not match the currently-pending prompt. The game advanced before the submission landed. Fetch the latest `turn_prompt` via the SSE stream and retry with the new `prompt_id`. Phase 2.5.",
            "enum": [
              "StaleTick"
            ],
            "type": "string"
          },
          {
            "description": "`POST .../actions` supplied an `action_index` outside the range of the pending prompt's `legal_actions`. Phase 2.5.",
            "enum": [
              "IllegalActionIndex"
            ],
            "type": "string"
          },
          {
            "description": "`POST .../actions` arrived with no prompt pending for this seat — either the game has not reached the seat yet or the last prompt was already satisfied. Phase 2.5.",
            "enum": [
              "NotYourTurn"
            ],
            "type": "string"
          },
          {
            "description": "`POST .../actions` targeted a seat that is not backed by an `http-remote` agent. AI-only seats cannot be driven via HTTP. Phase 2.5.",
            "enum": [
              "NoRemoteAgentAtSeat"
            ],
            "type": "string"
          },
          {
            "description": "A `POST /api/runs` payload named an agent kind that is CLI-only in the current phase — today `llm` and `stdio`. Server callers must use `playtest play --agents ...` from the command line. Phase 3.",
            "enum": [
              "AgentKindNotAllowedHere"
            ],
            "type": "string"
          }
        ]
      },
...
      "CreateRunRequest": {
        "description": "Request body for `POST /runs`.\n\n`config` is carried as untyped JSON so this crate stays game-agnostic; the server parses it against the named game's schema before accepting the run.",
        "properties": {
          "agents": {
            "description": "Ordered list of agent kind ids, one per seat. Length must fall within the target game's declared player-count range (e.g. exactly 2 for Cribbage, 2-4 for ShipWreck).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "config": {
            "description": "Game-specific configuration blob. Shape is defined by the target game's `config_schema` (see [`crate::registry::GameRegistryEntry`]).",
            "nullable": true
          },
          "game": {
            "description": "Registered game id (e.g. `\"cribbage\"`).",
            "type": "string"
          },
          "games_count": {
            "description": "How many games to play in this run. Must be >= 1.",
            "format": "uint32",
            "minimum": 0.0,
            "type": "integer"
          },
          "seed": {
            "description": "Optional base seed. `None` means \"pick one at run start\".",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "agents",
          "game",
          "games_count"
        ],
        "type": "object"
      },
...
      "RunStatus": {
        "description": "Lifecycle state of a run.",
        "oneOf": [
          {
            "description": "Accepted by the server but not yet started.",
            "enum": [
              "Pending"
            ],
            "type": "string"
          },
          {
            "description": "Currently executing games.",
            "enum": [
              "Running"
            ],
            "type": "string"
          },
          {
            "description": "All games finished successfully.",
            "enum": [
              "Completed"
            ],
            "type": "string"
          },
          {
            "description": "Terminated early due to an error.",
            "enum": [
              "Failed"
            ],
            "type": "string"
          }
        ]
      },
      "RunSummary": {
        "description": "Summary row returned by `GET /runs` and `GET /runs/:id`.",
        "properties": {
          "agents": {
            "description": "Agent kind ids, one per seat.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "created_at": {
            "description": "Wall-clock time the run was created, Unix epoch milliseconds.",
            "format": "uint64",
            "minimum": 0.0,
            "type": "integer"
          },
          "finished_at": {
            "description": "Wall-clock time the run finished, Unix epoch milliseconds. `None` while `status` is `Pending` or `Running`.",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true,
            "type": "integer"
          },
          "game": {
            "description": "Registered game id this run is playing.",
            "type": "string"
          },
          "games_completed": {
            "description": "Games that have finished (successfully or not) so far.",
            "format": "uint32",
            "minimum": 0.0,
            "type": "integer"
          },
          "games_count": {
            "description": "Total games requested.",
            "format": "uint32",
            "minimum": 0.0,
            "type": "integer"
          },
          "id": {
            "description": "Server-assigned run id.",
            "type": "string"
          },
          "seed": {
            "description": "Base seed chosen for the run.",
            "format": "uint64",
            "minimum": 0.0,
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "description": "Current lifecycle state."
          }
        },
        "required": [
          "agents",
          "created_at",
          "game",
          "games_completed",
          "games_count",
          "id",
          "seed",
          "status"
        ],
        "type": "object"
      },
      "SseFrame": {
        "description": "One frame in an SSE stream.\n\nSerialized with `#[serde(tag = \"kind\", content = \"data\")]`, so the JSON shape is `{\"kind\":\"event\",\"data\":{...}}` / `{\"kind\":\"heartbeat\"}` / etc.",
        "oneOf": [
          {
            "description": "First frame on every stream. `data` is the JSON object for the log's header record.",
            "properties": {
              "data": true,
              "kind": {
                "enum": [
                  "header"
                ],
                "type": "string"
              }
            },
            "required": [
              "data",
              "kind"
            ],
            "type": "object"
          },
          {
            "description": "An in-progress game event. `data` is the JSON object for one `Event` record from the log.",
            "properties": {
              "data": true,
              "kind": {
                "enum": [
                  "event"
                ],
                "type": "string"
              }
            },
            "required": [
              "data",
              "kind"
            ],
            "type": "object"
          },
          {
            "description": "Last frame when a game ends cleanly. `data` is the JSON object for the log's `Final` record.",
            "properties": {
              "data": true,
              "kind": {
                "enum": [
                  "final"
                ],
                "type": "string"
              }
            },
            "required": [
              "data",
              "kind"
            ],
            "type": "object"
          },
          {
            "description": "A remote agent is waiting for a submission. Ephemeral — not in the JSONL log, not resumable via `Last-Event-ID`. On reconnect the server re-emits this frame from the coordinator's pending state when the game is still waiting.",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/TurnPromptPayload"
              },
              "kind": {
                "enum": [
                  "turn_prompt"
                ],
                "type": "string"
              }
            },
            "required": [
              "data",
              "kind"
            ],
            "type": "object"
          },
          {
            "description": "Keep-alive tick so proxies do not time out idle streams. Carries no payload.",
            "properties": {
              "kind": {
                "enum": [
                  "heartbeat"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          }
        ]
      },
...
      "TurnPromptPayload": {
        "description": "Payload carried by [`SseFrame::TurnPrompt`]. A remote agent at `seat` is waiting for a submission; the client should POST an action_index into the legal list. Echo `prompt_id` back so the server can reject stale submissions.",
        "properties": {
          "legal_actions": {
            "description": "Legal actions indexed 0..N, each serialized as the game's `Action` type. The client picks one index and submits it.",
            "items": true,
            "type": "array"
          },
          "prompt_id": {
            "description": "Per-game monotonic id; echo on submission so the server can reject stale prompts (e.g., the game advanced before the client submitted).",
            "format": "uint64",
            "minimum": 0.0,
            "type": "integer"
          },
          "seat": {
            "description": "Zero-based seat index the prompt is addressed to.",
            "format": "uint8",
            "minimum": 0.0,
            "type": "integer"
          }
        },
        "required": [
          "legal_actions",
          "prompt_id",
          "seat"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Localhost HTTP + Server-Sent Events surface for the playtester engine. See docs/api-contract.md for versioning policy, SSE semantics, and error catalog.",
    "title": "Playtester API",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/RunSummary"
                      },
                      "type": "array"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Run summaries, newest first."
          }
        },
        "summary": "List active and recently-completed runs."
      },
      "post": {
        "description": "The run is accepted synchronously and executed in the background; poll `GET /api/runs/{run_id}` or subscribe to `/api/runs/{run_id}/stream` to observe progress.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RunSummary"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Run accepted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation failure (unknown game, unknown agent, bad `games_count`, wrong agent count)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal failure registering the run."
          }
        },
        "summary": "Create a run and start playing games."
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "parameters": [
          {
            "description": "UUID of the run.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RunSummary"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Run summary."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "`run_id` is not a known run."
          }
        },
        "summary": "Fetch one run's summary."
      }
    },
...
    "/api/runs/{run_id}/stream": {
      "get": {
        "description": "Server-Sent Events stream. Frames fire when games inside this run start and finish, plus one terminal `run_complete` frame. Clients may reconnect with `Last-Event-ID` but run-level frames are not re-delivered — use the per-game stream for replay semantics.",
        "parameters": [
          {
            "description": "UUID of the run.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseFrame"
                }
              }
            },
            "description": "Infinite SSE stream of `{game_started, game_finished, run_complete}` frames."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "items": {
                        "$ref": "#/components/schemas/ApiError"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "api_version",
                    "errors"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "`run_id` is not a known run."
          }
        },
        "summary": "SSE stream of run-level events."
      }
    }
  }
}